Remove unused locales (#1)
* [misc] release process fixes (#4961)

* Fix locale abbreviation in comment

* Update package-lock.json

* [misc] Fix type of WeekSpec (#4634)

* Fix "the the" typo in test (#4578)

* [locale] Fix #4568: Insert punjabi for "next" (#4583)

* Fix type of WeekSpec

You can set `dow` without setting `doy`, so the type should reflect that

* [new locale] (oc-lnc) Occitan (#4957)

* Add oc-lnc files

* Fix calendar test

* Fix the rest of tests

* Add locale author

* Dont make changes to package-lock

* Add newline at end of file

* [bugfix] Fix #4962 updateLocale with null to reset (#4966)

* [locale] es: Add es translation for invalidDate (#4560)

* Add es translation for invalidDate

* Fix lint

* [locale] improve mk translation (#4958)

* Update mk.js

* Update mk.js

* Update mk.js

* Update mk.js

* [locale] improve mk translation

* [tests] Fix various text typos (#4976)

* [locale] mr (Marathi) (#4990)

* tests according to changes in mr.js

* [locale] mr (Marathi)

We Marathi people follow standards mentioned below, to label timespans.

Label | Time Span 
------------ | -------------
पहाटे (post midnight, pre-sunrise) |    >= 12am to < 6am 
सकाळी (post sunrise, pre-noon) |    >= 6am to < 12pm
दुपारी (afternoon) |    >= 12pm to < 5pm
सायंकाळी (evening) |    >= 5pm to < 8pm
रात्री (night) |    >= 8pm to < 12am

Actually, 12am is confusing in Marathi tradition. In modern clock system, Moment 12am is considered in arriving date as 0am. And in Marathi tradition it is referred to as final moment of the previous date (Sort of, 0:01 is starting minute of the arriving date) . To comply with modern clock system, I propose to refer to 0am as पहाटे , so that it is considered as moment in the arriving date. Hope this discription itself is not confusing :)

Another point is about range of पहाटे, ie. >= 0am to < 6am. Though conventionally, upto 2 am, we refer to the time as रात्री or पहाटे, both alternatively, I want to specifically call पहाटे for early times in the arriving date and रात्री for late times in the departing date, to make it more specific and well-defined, avoiding confusion about date of the time.

* [locale] correct translation of locale for Vietnam (#4874)

* correct translation of locale for Vietnam

* correct translation of locale for Vietnam (edit proper src files)

* Update author comment line

* [locale] (en-sg) Rename locale en-SG to en-sg (#5025)

* [locale] typos in Irish locale for months (#5127)

Fixed some typos in the Irish locale regarding months.

* [locale] (br) Breton move from 12-hour to 24-hour notation (#4975)

* Breton : move from 12-hour to 24-hour notation #4974 https://github.com/moment/moment/issues/4974

* Replace [e] by :

* [locale] bn: Improve translation as per Bangla Academy (#5055)

* Delete unused locales

* Remove unused locales

* Update test and build

* Add untracked locale